Adds generated moon graphics, dates & visibiliy values
The moon phase visuals are produced by essentially sliding a circle
horizontally, while making sure its radius is large enough that it
has an intersection at the top and bottom of a smaller circle
of fixed radius and position.

The dates for each moon phase are calculated and formatted inside
the GetLunarData() function. This should probably be refactored to
use the julian days field to calculate the current date.