Graphics update (#6)
* Adds generated moon graphics, dates & visibiliy values

The moon phase visuals are produced by essentially sliding a circle
horizontally, while making sure its radius is large enough that it
has an intersection at the top and bottom of a smaller circle
of fixed radius and position.

The dates for each moon phase are calculated and formatted inside
the GetLunarData() function. This should probably be refactored to
use the julian days field to calculate the current date.

* Fixed incorrect generation of waning moon phase

* Fix SIGFPE bug on raspberry pi

This is a very strange bug thats technically been around
for a while. It only manifested when I changed the class
definition for LunarData, and even then it only materialized
on the raspberry pi build.

Co-authored-by: Bob <atomicsteamend@gmail.com>